+++ api/stickers.js
